integrate weather in clock
