+++ immichFrame.Web/src/lib/components/home-page/home-page.svelte
@@ -10,7 +10,6 @@
 	import ImageComponent from '../elements/image-component.svelte';
 	import { configStore } from '$lib/stores/config.store';
 	import ErrorElement from '../elements/error-element.svelte';
-	import Weather from '../elements/weather.svelte';
 
 	let imageData: Blob | null;
 	let assetData: api.AssetResponseDto | null;
@@ -92,8 +91,6 @@
 			{assetData}
 			{imageData}
 		/>
-
-		<Weather showWeather={true} />
 
 		<OverlayControls
 			on:next={async () => {
